+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/internal/SyncConfigurationImpl.kt
@@ -17,18 +17,31 @@
 package io.realm.kotlin.mongodb.internal
 
 import io.realm.kotlin.internal.InternalConfiguration
+import io.realm.kotlin.internal.MutableLiveRealmImpl
 import io.realm.kotlin.internal.RealmImpl
+import io.realm.kotlin.internal.TypedFrozenRealmImpl
+import io.realm.kotlin.internal.interop.FrozenRealmPointer
+import io.realm.kotlin.internal.interop.LiveRealmPointer
+import io.realm.kotlin.internal.interop.RealmAppPointer
 import io.realm.kotlin.internal.interop.RealmConfigurationPointer
 import io.realm.kotlin.internal.interop.RealmInterop
+import io.realm.kotlin.internal.interop.RealmSyncConfigurationPointer
 import io.realm.kotlin.internal.interop.RealmSyncSessionPointer
+import io.realm.kotlin.internal.interop.SyncAfterClientResetHandler
+import io.realm.kotlin.internal.interop.SyncBeforeClientResetHandler
 import io.realm.kotlin.internal.interop.SyncErrorCallback
 import io.realm.kotlin.internal.interop.sync.PartitionValue
 import io.realm.kotlin.internal.interop.sync.SyncError
+import io.realm.kotlin.internal.interop.sync.SyncSessionResyncMode
 import io.realm.kotlin.internal.platform.freeze
+import io.realm.kotlin.mongodb.exceptions.ClientResetRequiredException
 import io.realm.kotlin.mongodb.exceptions.DownloadingRealmTimeOutException
 import io.realm.kotlin.mongodb.subscriptions
+import io.realm.kotlin.mongodb.sync.DiscardUnsyncedChangesStrategy
 import io.realm.kotlin.mongodb.sync.InitialRemoteDataConfiguration
 import io.realm.kotlin.mongodb.sync.InitialSubscriptionsConfiguration
+import io.realm.kotlin.mongodb.sync.ManuallyRecoverUnsyncedChangesStrategy
+import io.realm.kotlin.mongodb.sync.SyncClientResetStrategy
 import io.realm.kotlin.mongodb.sync.SyncConfiguration
 import io.realm.kotlin.mongodb.sync.SyncMode
 import io.realm.kotlin.mongodb.sync.SyncSession
@@ -40,6 +53,7 @@ internal class SyncConfigurationImpl(
     internal val partitionValue: PartitionValue?,
     override val user: UserImpl,
     override val errorHandler: SyncSession.ErrorHandler,
+    override val syncClientResetStrategy: SyncClientResetStrategy,
     override val initialSubscriptions: InitialSubscriptionsConfiguration?,
     override val initialRemoteData: InitialRemoteDataConfiguration?
 ) : InternalConfiguration by configuration, SyncConfiguration {
@@ -68,33 +82,159 @@ internal class SyncConfigurationImpl(
     }
 
     override fun createNativeConfiguration(): RealmConfigurationPointer {
-        val ptr = configuration.createNativeConfiguration()
+        val ptr: RealmConfigurationPointer = configuration.createNativeConfiguration()
         return syncInitializer(ptr)
     }
 
     private val syncInitializer: (RealmConfigurationPointer) -> RealmConfigurationPointer
 
+    private interface ClientResetStrategyHelper {
+        fun initialize(nativeSyncConfig: RealmSyncConfigurationPointer)
+        fun onSyncError(session: SyncSession, appPointer: RealmAppPointer, error: SyncError)
+    }
+
+    private class DiscardUnsyncedChangesHelper constructor(
+        val strategy: DiscardUnsyncedChangesStrategy,
+        val configuration: InternalConfiguration
+    ) : ClientResetStrategyHelper {
+        override fun initialize(nativeSyncConfig: RealmSyncConfigurationPointer) {
+            RealmInterop.realm_sync_config_set_resync_mode(
+                nativeSyncConfig,
+                SyncSessionResyncMode.RLM_SYNC_SESSION_RESYNC_MODE_DISCARD_LOCAL
+            )
+
+            val onBefore: SyncBeforeClientResetHandler = object : SyncBeforeClientResetHandler {
+                override fun onBeforeReset(realmBefore: FrozenRealmPointer) {
+                    strategy.onBeforeReset(TypedFrozenRealmImpl(realmBefore, configuration))
+                }
+            }
+
+            RealmInterop.realm_sync_config_set_before_client_reset_handler(
+                nativeSyncConfig,
+                onBefore
+            )
+
+            val onAfter: SyncAfterClientResetHandler = object : SyncAfterClientResetHandler {
+                override fun onAfterReset(
+                    realmBefore: FrozenRealmPointer,
+                    realmAfter: LiveRealmPointer,
+                    didRecover: Boolean
+                ) {
+                    // Needed to allow writes on the Mutable after Realm
+                    RealmInterop.realm_begin_write(realmAfter)
+
+                    @Suppress("TooGenericExceptionCaught")
+                    try {
+                        strategy.onAfterReset(
+                            TypedFrozenRealmImpl(realmBefore, configuration),
+                            MutableLiveRealmImpl(realmAfter, configuration)
+                        )
+
+                        // Callback completed successfully we can safely commit the changes
+                        // user might have cancelled the transaction manually
+                        if (RealmInterop.realm_is_in_transaction(realmAfter)) {
+                            RealmInterop.realm_commit(realmAfter)
+                        }
+                    } catch (exception: Throwable) {
+                        // Cancel the transaction
+                        // user might have cancelled the transaction manually
+                        if (RealmInterop.realm_is_in_transaction(realmAfter)) {
+                            RealmInterop.realm_rollback(realmAfter)
+                        }
+                        // Rethrow so core can send it over again
+                        throw exception
+                    }
+                }
+            }
+
+            RealmInterop.realm_sync_config_set_after_client_reset_handler(
+                nativeSyncConfig,
+                onAfter
+            )
+        }
+
+        override fun onSyncError(
+            session: SyncSession,
+            appPointer: RealmAppPointer,
+            error: SyncError
+        ) {
+            // If there is a user exception we appoint it as the cause of the client reset
+            strategy.onError(
+                session,
+                ClientResetRequiredException(appPointer, error)
+            )
+        }
+    }
+
+    private class ManuallyRecoverUnsyncedChangesHelper(
+        val strategy: ManuallyRecoverUnsyncedChangesStrategy
+    ) : ClientResetStrategyHelper {
+        override fun initialize(nativeSyncConfig: RealmSyncConfigurationPointer) {
+            RealmInterop.realm_sync_config_set_resync_mode(
+                nativeSyncConfig,
+                SyncSessionResyncMode.RLM_SYNC_SESSION_RESYNC_MODE_MANUAL
+            )
+        }
+
+        override fun onSyncError(
+            session: SyncSession,
+            appPointer: RealmAppPointer,
+            error: SyncError
+        ) {
+            strategy.onClientReset(
+                session,
+                ClientResetRequiredException(appPointer, error)
+            )
+        }
+    }
+
     init {
         // We need to freeze `errorHandler` reference on initial thread
         val userErrorHandler = errorHandler
-        val errorCallback = object : SyncErrorCallback {
-            override fun onSyncError(pointer: RealmSyncSessionPointer, error: SyncError) {
-                userErrorHandler.onError(SyncSessionImpl(pointer), convertSyncError(error))
-            }
-        }.freeze()
+        val resetStrategy = syncClientResetStrategy.freeze()
+        val frozenAppPointer = user.app.nativePointer.freeze()
+
+        val initializerHelper = when (resetStrategy) {
+            is DiscardUnsyncedChangesStrategy ->
+                DiscardUnsyncedChangesHelper(resetStrategy, configuration)
+            is ManuallyRecoverUnsyncedChangesStrategy ->
+                ManuallyRecoverUnsyncedChangesHelper(resetStrategy)
+            else -> throw IllegalArgumentException("Unsupported client reset strategy: $resetStrategy")
+        }
+
+        val errorCallback =
+            SyncErrorCallback { pointer: RealmSyncSessionPointer, error: SyncError ->
+                val session = SyncSessionImpl(pointer)
+                val syncError = convertSyncError(error)
+
+                // Notify before/after callbacks too if error is client reset
+                if (error.isClientResetRequested) {
+                    initializerHelper.onSyncError(session, frozenAppPointer, error)
+                } else {
+                    userErrorHandler.onError(session, syncError)
+                }
+            }.freeze()
 
         syncInitializer = { nativeConfig: RealmConfigurationPointer ->
-            val nativeSyncConfig = if (partitionValue == null) {
+            val nativeSyncConfig: RealmSyncConfigurationPointer = if (partitionValue == null) {
                 RealmInterop.realm_flx_sync_config_new(user.nativePointer)
             } else {
-                RealmInterop.realm_sync_config_new(user.nativePointer, partitionValue.asSyncPartition())
+                RealmInterop.realm_sync_config_new(
+                    user.nativePointer,
+                    partitionValue.asSyncPartition()
+                )
             }
 
             RealmInterop.realm_sync_config_set_error_handler(
                 nativeSyncConfig,
                 errorCallback
             )
+
+            // Do any initialization required for the strategies
+            initializerHelper.initialize(nativeSyncConfig)
+
             RealmInterop.realm_config_set_sync_config(nativeConfig, nativeSyncConfig)
+
             nativeConfig
         }
     }